+++ packages/packages/core/editor-interactions/src/components/interactions-list.tsx
@@ -1,13 +1,14 @@
 import * as React from 'react';
-import { createContext, useCallback, useContext, useEffect, useMemo, useRef } from 'react';
+import { useCallback, useEffect, useMemo, useRef } from 'react';
 import { Repeater } from '@elementor/editor-controls';
 import { InfoCircleFilledIcon, PlayerPlayIcon } from '@elementor/icons';
 import { Alert, AlertTitle, Box, IconButton, Tooltip } from '@elementor/ui';
 import { __ } from '@wordpress/i18n';
 
 import type { ElementInteractions, InteractionItemPropValue, InteractionItemValue } from '../types';
+import { InteractionItemContextProvider } from '../contexts/interactions-item-context';
 import { buildDisplayLabel, createDefaultInteractionItem, extractString } from '../utils/prop-value-utils';
-import { InteractionDetails } from './interaction-details';
+import { InteractionsListItem } from './interactions-list-item';
 export const MAX_NUMBER_OF_INTERACTIONS = 5;
 
 export type InteractionListProps = {
@@ -16,13 +17,6 @@ export type InteractionListProps = {
 	onPlayInteraction: ( interactionId: string ) => void;
 	triggerCreateOnShowEmpty?: boolean;
 };
-
-type InteractionItemContextValue = {
-	onInteractionChange: ( index: number, newInteractionValue: InteractionItemValue ) => void;
-	onPlayInteraction: ( interactionId: string ) => void;
-};
-
-const InteractionItemContext = createContext< InteractionItemContextValue | null >( null );
 
 export function InteractionsList( props: InteractionListProps ) {
 	const { interactions, onSelectInteractions, onPlayInteraction, triggerCreateOnShowEmpty } = props;
@@ -101,7 +95,7 @@ export function InteractionsList( props: InteractionListProps ) {
 	);
 
 	return (
-		<InteractionItemContext.Provider value={ contextValue }>
+		<InteractionItemContextProvider value={ contextValue }>
 			<Repeater
 				openOnAdd
 				openItem={ triggerCreateOnShowEmpty ? 0 : undefined }
@@ -117,7 +111,7 @@ export function InteractionsList( props: InteractionListProps ) {
 					initialValues: createDefaultInteractionItem(),
 					Label: ( { value }: { value: InteractionItemPropValue } ) => buildDisplayLabel( value.value ),
 					Icon: () => null,
-					Content,
+					Content: InteractionsListItem,
 					actions: ( value: InteractionItemPropValue ) => (
 						<Tooltip key="preview" placement="top" title={ __( 'Preview', 'elementor' ) }>
 							<IconButton
@@ -131,35 +125,6 @@ export function InteractionsList( props: InteractionListProps ) {
 					),
 				} }
 			/>
-		</InteractionItemContext.Provider>
+		</InteractionItemContextProvider>
 	);
 }
-
-const Content = ( { index, value }: { index: number; value: InteractionItemPropValue } ) => {
-	const context = useContext( InteractionItemContext );
-
-	const handleChange = useCallback(
-		( newInteractionValue: InteractionItemValue ) => {
-			context?.onInteractionChange( index, newInteractionValue );
-		},
-		[ context, index ]
-	);
-
-	const handlePlayInteraction = useCallback(
-		( interactionId: string ) => {
-			context?.onPlayInteraction( interactionId );
-		},
-		[ context ]
-	);
-
-	return (
-		<>
-			<InteractionDetails
-				key={ index }
-				interaction={ value.value }
-				onChange={ handleChange }
-				onPlayInteraction={ handlePlayInteraction }
-			/>
-		</>
-	);
-};
